Rename `cut_wire_to_move` to `wire_cutting_transforms`
