+++ circuit_knitting/cutting/__init__.py
@@ -88,7 +88,7 @@ from .cutting_decomposition import (
 )
 from .cutting_evaluation import execute_experiments, CuttingExperimentResults
 from .cutting_reconstruction import reconstruct_expectation_values
-from .cut_wire_to_move import cut_wires, expand_observables
+from .wire_cutting_transforms import cut_wires, expand_observables
 
 __all__ = [
     "partition_circuit_qubits",
